requirements.txt
cffi==1.14.2
click==7.1.2
cryptography==3.0
Flask==1.1.2
Flask-API==2.0
Flask-Cors==3.0.8
Flask-Limiter==1.4
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
limits==1.5.1
MarkupSafe==1.1.1
Pillow==7.2.0
pkg-resources==0.0.0
pycparser==2.20
pyOpenSSL==19.1.0
pyserial==3.4
RPi.GPIO==0.7.0
six==1.15.0
SQLAlchemy==1.3.19
thermalprinter==0.2.0
Werkzeug==1.0.1
uwsgi